+++ modules/client/test/integration/tokenVoting-client/methods.test.ts
@@ -348,7 +348,7 @@ describe("Token Voting Client", () => {
         const ctxPlugin = ContextPlugin.fromContext(ctx);
         const client = new TokenVotingClient(ctxPlugin);
         const limit = 5;
-        const status = ProposalStatus.EXECUTED;
+        const status = ProposalStatus.DEFEATED;
         const params: IProposalQueryParams = {
           limit,
           sortBy: ProposalSortBy.CREATED_AT,